pyproject.toml
[tool.poetry]
name = "pyCirclize"
version = "0.1.0"
description = "Circular visualization in Python"
authors = ["moshi4"]
license = "MIT"
homepage = "https://moshi4.github.io/pyCirclize/"
repository = "https://github.com/moshi4/pyCirclize/"
readme = "README.md"
keywords = [
    "matplotlib",
    "visualization",
    "bioinformatics",
    "circos",
    "chord-diagram",
]
classifiers = [
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Framework :: Matplotlib",
]
include = ["tests"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov=src --tb=line --cov-report=xml --cov-report=term"
testpaths = ["tests"]

[tool.poetry.scripts]

[tool.poetry.dependencies]
python = "^3.8"
matplotlib = "^3.5.2"
biopython = "^1.79"
numpy = "^1.21.1"
pandas = "^1.3.5"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.10.0"
flake8 = "^4.0.1"
pydocstyle = "^6.1.1"
pytest-cov = "^4.0.0"
ipykernel = "^6.13.0"
mkdocs = "^1.2"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}
mkdocs-jupyter = "^0.21.0"
mkdocs-material = "^8.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
